+++ Sources/Instructions/Protocols and Enums/Public/Delegates/CoachMarksControllerDelegate.swift
@@ -34,6 +34,9 @@ public protocol CoachMarksControllerDelegate: AnyObject {
 
     func shouldHandleOverlayTap(in coachMarksController: CoachMarksController,
                                 at index: Int) -> Bool
+    
+    func coachMarksController(_ coachMarksController: CoachMarksController,
+                              didTapCoachMarkAt index: Int)
 }
 
 public extension CoachMarksControllerDelegate {
@@ -80,6 +83,9 @@ public extension CoachMarksControllerDelegate {
                                 at index: Int) -> Bool {
         return true
     }
+
+    func coachMarksController(_ coachMarksController: CoachMarksController,
+                              didTapCoachMarkAt index: Int) { }
 }
 
 protocol CoachMarksControllerProxyDelegate: AnyObject {
@@ -103,4 +109,6 @@ protocol CoachMarksControllerProxyDelegate: AnyObject {
     func didEndShowingBySkipping(_ skipped: Bool)
 
     func shouldHandleOverlayTap(at index: Int) -> Bool
+
+    func didTapCoachMark(at index: Int)
 }